+++ NEWS.rst
@@ -5,6 +5,13 @@ Changes in 5.11 (released ??/??/2014)
 
 *	:func:`misc.javaexpr` now supports sets.
 
+*	Sisyphus jobs now have a new method :meth:`tasks` that loops over an iterable
+	and calls :meth:`task` for each item::
+
+		items = sys.modules.items()
+		for (name, module) in self.tasks(items, "module", lambda kv: kv[0]):
+			self.log("module is {}".format(module))
+
 
 Changes in 5.10 (released 10/09/2014)
 -------------------------------------